# services/go-auth/docs/swagger.yaml
basePath: /
definitions:
  models.UserCreateReq:
    description: Данные для регистрации или входа пользователя
    properties:
      login:
        example: user123
        maxLength: 20
        minLength: 3
        type: string
      password:
        example: strongPassword123
        maxLength: 32
        minLength: 6
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Authentication service API
  title: Go Auth API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Проверяет учетные данные и возвращает JWT токены в cookies
      parameters:
      - description: Учетные данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UserCreateReq'
      produces:
      - application/json
      responses:
        "204":
          description: Успешная аутентификация, токены установлены в cookies
        "400":
          description: Неверный запрос или неверные логин/пароль
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Аутентификация пользователя
      tags:
      - Аутентификация
  /logout:
    get:
      description: Удаляет JWT токены из cookies и хранилища
      produces:
      - application/json
      responses:
        "204":
          description: Успешный выход, токены удалены
        "400":
          description: Неверный запрос
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Выход из системы
      tags:
      - Аутентификация
  /register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя и возвращает JWT токены в cookies
      parameters:
      - description: Данные для регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UserCreateReq'
      produces:
      - application/json
      responses:
        "204":
          description: Успешная регистрация, токены установлены в cookies
        "400":
          description: Неверный запрос или неверные логин/пароль
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Регистрация нового пользователя
      tags:
      - Аутентификация
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"